everything but the scoreboard
the game works pretty well. Havent tested latest change to speed increases yet. Im so close to finishing that I want to finish all of it and test a final copy to what im sure will be a broken game. running the file shows a black page but hitting where the play button "should be" activates it.